@@ -1,11 +1,4 @@
-import Vue, { VNode } from 'vue'
+import { createApp } from 'vue'
 import Dev from './serve.vue'
-import VueCompositionApi from '@vue/composition-api'
 
-Vue.config.productionTip = false
-
-Vue.use(VueCompositionApi)
-
-new Vue({
-  render: (h): VNode => h(Dev),
-}).$mount('#app')
+createApp(Dev).mount('#app')
