@@ -1,4 +1,4 @@
-const devPresets = ['@vue/babel-preset-app']
+const devPresets = ['@vue/cli-plugin-babel/preset']
 const buildPresets = ['@babel/preset-env', '@babel/preset-typescript']
 module.exports = {
   presets: process.env.NODE_ENV === 'development' ? devPresets : buildPresets,
