@@ -3,7 +3,7 @@
 </template>
 
 <script lang="ts">
-import { defineComponent } from '@vue/composition-api'
+import { defineComponent } from 'vue'
 import { StackedBar } from 'rough-viz'
 import { stackedBarChartOptions } from '@/shared'
 import { useSetupRoughVizChart } from '@/composables'
