@@ -1,7 +1,7 @@
 export { default as BarChart } from './BarChart.vue'
-export { default as DonutChart } from './DonutChart.vue'
-export { default as HorizontalBarChart } from './HorizontalBarChart.vue'
-export { default as LineChart } from './LineChart.vue'
-export { default as PieChart } from './PieChart.vue'
-export { default as ScatterChart } from './ScatterChart.vue'
-export { default as StackedBarChart } from './StackedBarChart.vue'
+// export { default as DonutChart } from './DonutChart.vue'
+// export { default as HorizontalBarChart } from './HorizontalBarChart.vue'
+// export { default as LineChart } from './LineChart.vue'
+// export { default as PieChart } from './PieChart.vue'
+// export { default as ScatterChart } from './ScatterChart.vue'
+// export { default as StackedBarChart } from './StackedBarChart.vue'
